float

+++ log.capnp
@@ -1357,7 +1357,7 @@ struct UploaderState {
   rawQueueCount @3 :UInt32;
 
   # stats for last successfully uploaded file
-  lastTime @4 :UInt32;   # s
+  lastTime @4 :Float32;  # s
   lastSpeed @5 :Float32; # MB/s
   lastFilename @6 :Text;
 }
